requirements.txt
# EduShield - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Machine Learning
scikit-learn==1.3.2
imbalanced-learn==0.11.0
pandas==2.1.3
numpy==1.26.2

# Data Validation
pydantic==2.5.0
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4